Start refactoring transcript rendering into band_track
